# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /user:
    get:
      operationId: UserController_findAll
      summary: Listar todos os usuários
      parameters: []
      responses:
        "200":
          description: ""
      tags: &a4
        - Usuário
      security:
        - bearer: []
    post:
      operationId: UserController_create
      summary: Criar um novo usuário
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
      tags: *a4
      security:
        - bearer: []
    patch:
      operationId: UserController_update
      summary: Atualizar informações do usuário
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
      tags: *a4
      security:
        - bearer: []
  /user/me:
    get:
      operationId: UserController_findMe
      summary: Buscar informações do usuário logado
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
      security:
        - bearer: []
  /user/forget-password:
    post:
      operationId: UserController_createRecoveryCode
      summary: Rota para gerar código de recuperação de senha
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgetPasswordDto"
      responses:
        "201":
          description: ""
      tags: *a4
  /user/forget-password/check-code:
    post:
      operationId: UserController_checkCode
      summary: Rota para gerar código de recuperação de senha
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckCodeDto"
      responses:
        "201":
          description: ""
      tags: *a4
  /user/forget-password/code:
    post:
      operationId: UserController_forgetPasswordWithCode
      summary: Rota para gerar código de recuperação de senha
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgetPasswordWithCodeDto"
      responses:
        "201":
          description: ""
      tags: *a4
  /user/change-password/old-password:
    post:
      operationId: UserController_changePassword
      summary: Alterar senha (com senha antiga)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordWithOldPasswordDto"
      responses:
        "201":
          description: ""
      tags: *a4
      security:
        - bearer: []
  /user/change-password/token:
    post:
      operationId: UserController_changePasswordWithToken
      summary: Alterar senha (com token)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordWithTokenDto"
      responses:
        "201":
          description: ""
      tags: *a4
  /auth/register:
    post:
      operationId: AuthController_create
      summary: Rota para criar um novo usuário
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
      tags: &a5
        - Auth
  /auth/login:
    post:
      operationId: AuthController_login
      summary: Rota para realizar o login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserDto"
      responses:
        "201":
          description: ""
      tags: *a5
...
info:
  title: Zapdiviser
  description: API para o Zapdiviser
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateUserDto:
      type: object
      properties: {}
    ForgetPasswordDto:
      type: object
      properties: {}
    CheckCodeDto:
      type: object
      properties: {}
    ForgetPasswordWithCodeDto:
      type: object
      properties: {}
    UpdateUserDto:
      type: object
      properties: {}
    UpdatePasswordWithOldPasswordDto:
      type: object
      properties: {}
    UpdatePasswordWithTokenDto:
      type: object
      properties: {}
    LoginUserDto:
      type: object
      properties: {}
...